chore: move BebopOrderType to models.rs
