+++ src/encoding/evm/swap_encoder/mod.rs
@@ -1,5 +1,3 @@
 mod builder;
 pub mod swap_encoder_registry;
 mod swap_encoders;
-
-pub use swap_encoders::BebopOrderType;
